one big function is divided into small ones + the remove_extracted_applications() function has been added

+++ src/main.py
@@ -45,71 +45,77 @@ SILENT_INSTALL_APPS = {
 
 
 # функция для скачивания архива
-def download_archive(url, save_path, update_status_callback):
-    response = requests.get(url, stream=True)
-    total_length = response.headers.get('content-length')
+def download_archive(url, save_path, selection, update_status_callback):
+    if not os.path.exists(ARCHIVE_PATH):
+        update_status_callback("Ошибка: архив не найден")
+        time.sleep(1)
 
-    if total_length is None:
-        with open(save_path, 'wb') as file:
-            file.write(response.content)
-        update_status_callback("Скачивание завершено")
-    else:
-        dl = 0
-        total_length = int(total_length)
-        with open(save_path, 'wb') as file:
-            for data in response.iter_content(chunk_size=4096):
-                dl += len(data)
-                file.write(data)
-                done = int(100 * dl / total_length)
-                update_status_callback(f"Скачивается... {done}%")
-        update_status_callback("Скачивание завершено")
+        response = requests.get(url, stream=True)
+        total_length = response.headers.get('content-length')
+
+        if total_length is None:
+            with open(save_path, 'wb') as file:
+                file.write(response.content)
+            update_status_callback("Скачивание завершено")
+        else:
+            dl = 0
+            total_length = int(total_length)
+            with open(save_path, 'wb') as file:
+                for data in response.iter_content(chunk_size=4096):
+                    dl += len(data)
+                    file.write(data)
+                    done = int(100 * dl / total_length)
+                    update_status_callback(f"Скачивается... {done}%")
+            update_status_callback("Скачивание завершено")
+
+    extract_applications(url, save_path, selection, update_status_callback)
 
 
-# Функция извлечения
-def download_and_extract(url, save_path, selection, update_status_callback):
-
-    # Проверяем наличие архива
+def extract_applications(url, save_path, selection, update_status_callback):
     if not os.path.exists(ARCHIVE_PATH):
         update_status_callback("Ошибка: архив не найден")
         time.sleep(1)
         download_archive(url, save_path, update_status_callback)
     if os.path.exists(ARCHIVE_PATH):
-        # Извлекаем выбранные приложения
         with zipfile.ZipFile(ARCHIVE_PATH, "r") as archive:
             for app in selection:
                 if app in APPLICATION:
                     app_file = APPLICATION[app]
                     archive.extract(app_file, EXTRACT_PATH)
+                    update_status_callback(f"{app} извлечено")
                     print(f"Приложение {app} успешно извлечено!")
-        # install_extracted_programs(update_status_callback)
-                    update_status_callback(f"{app} распакован")
+                    time.sleep(0.2)
         time.sleep(1)
+    install_applications(selection, update_status_callback)
 
+
+def install_applications(selection, update_status_callback):
     silent_install_apps = [app for app in selection if app in SILENT_INSTALL_APPS]
     regular_install_apps = [app for app in selection if app not in SILENT_INSTALL_APPS]
 
-    # Установка тихих приложений
     for app in silent_install_apps:
         if app in APPLICATION:
             app_file = os.path.join(EXTRACT_PATH, APPLICATION[app])
             if app == "7-Zip":
-                update_status_callback(f"{app} устанвливается.")
+                update_status_callback(f"{app} устанавливается.")
                 subprocess.run([app_file, "/S"], check=True, shell=True)
             else:
-                update_status_callback(f"{app} устанвливается.")
+                update_status_callback(f"{app} устанавливается.")
                 subprocess.run([app_file, "/S", "/silent", "/verysilent", "/quiet", "/qn"], check=True, shell=True)
-            update_status_callback(f"{app} установлен.")
+            update_status_callback(f"{app} установлен тихо")
             print(f"Приложение {app} установлено тихо.")
             time.sleep(2)
 
-    # Установка остальных приложений
     for app in regular_install_apps:
         if app in APPLICATION:
             app_file = os.path.join(EXTRACT_PATH, APPLICATION[app])
             subprocess.run([app_file], check=True, shell=True)  # Обычная установка
-            print(f"Приложение {app} установлено.")
             update_status_callback(f"{app} установлен.")
+            print(f"Приложение {app} установлено.")
             time.sleep(2)
+
+    time.sleep(2)
+    remove_extracted_applications()
 
 
 def disable_autostart(update_status_callback):
@@ -133,6 +139,21 @@ def disable_autostart(update_status_callback):
         print(f"Ошибка при отключении автозагрузки: {e}")
         update_status_callback("Ошибка при отключении\nавтозагрузки")
         return False
+
+
+def remove_extracted_applications():
+    try:
+        if os.path.exists(EXTRACT_PATH):
+            for file_name in os.listdir(EXTRACT_PATH):
+                file_path = os.path.join(EXTRACT_PATH, file_name)
+                if os.path.isfile(file_path):
+                    os.remove(file_path)
+                    print(f"Удален файл: {file_name}")
+                    time.sleep(0.5)
+        else:
+            print("Папка не найдена")
+    except Exception as e:
+        print(f"Ошибка при удалении файлов: {e}")
 
 
 if __name__ == "__main__":
